service: change reconnect logic
today we invoked a groutine that would endlessly try to reconnect to
cluster.
it would never get new conn though cause it would never evaluate the new
connMap when updated.

current code will try to reconnect. if fail will start a timer that
would invoke reconnect in the future.

addtionally the same goroutine would trigger a reconnect on AEN as well
as updated cache which will solve the reconnect to stale connMap we have
today.