service, nvme-client: retry if client fails to connect nvme devices
We had some issues (in dual nodes tests) where client failed to connect
nvme devices. In those cases we got an error: "write /dev/nvme-fabrics: connection timed out".
This error was ignored and we could't find the block device of the target
until we tried to reconnect or create a new volume which triggered a new
connection.
In this commit we resolve this issue by reconnect once we get this error.

Issue: LBM1-27816

Signed-off-by: Guy Goldberg <guy@lightbitslabs.com>