Set hostid during nvme connection establishment
Starting from kernel v6.5-rc1, there is a new change regarding
hostid and hostnqn.
When hostnqn is passed without a specified hostid, the default
hostid generated by nvme during initialization is used.
The default hostid is already associated with the default hostnqn.
So, when attempting to connect with a specified hostnqn, the connection
fails due to encountering the same hostid (x) but different hostnqn (y).
Therefore, it is now required that hostid be provided with hostnqn as a pair.
This pairing ensures proper host identification in the updated kernel versions.

See the commit associated with the new change:
https://github.com/torvalds/linux/commit/ae8bd606e09bbdb2607c8249872cc2aeaf2fcc72

Issue: LBM1-30942

Signed-off-by: Guy Goldberg <guy@lightbitslabs.com>