service: when we close stopCh we need to nil it to avoid panic
in case we fail to connect and then we reconnect we would get
to a point where we close(stopCh) and on the second iteration
we would go into the if stopCh != nil { and try to close it
again.

this change will make sure we only close it once and will not
go in this if again.

this issue occured in situations where the DSC failed to connect
then succeeded and in the third iteration it would panic
on closing already closed chan.

issue: LBM1-33584