fix ctrl-loss-tmo flag support
till now we didn't support this flag in the DC configuraion.
it was added only on the connect command line but not as a service.
needed to update the parser of the file, update the cache
and pass this information to the ConnectAll command from
the service.

we differneciate between entries to the cache coming from user/referrals

the reason we do that is that a user may provide ctrl-loss-tmo on a server that is
not the default, and in case a referral will notify on a new server added we will
not have this value.

when we add the referral entry we will go over all the user entries and if
one of them is not nil we will use this one and apply if to the referral
one. sadly the DS will not be able to provide us this info about the new
server we connected.

issue: LBM1-35562