fail-fast of /dev/nvme-fabrics is not present.
there are still OS out that does not load nvme-tcp
by default. this will cause the DC to not function properly
when it tries to connect to the io-controller.

today we will see it only in the logs but the server will
run successfully and it is tricky to locate.

this change will fail-fast, on service load and will report
what in high probablility needed to be done: nvme-tcp.