Dockerfile.discovery-client: add curl to image
if we want to run liveness probes on discovery-client we need to
run it from the container.

when a service exposes an endpoint - like this one exposing http
it is good practice to expose an healthz endpoint that can be
invoked by k8s or docker-compose and indicate about container
status.