fix: enable DebugLevel trace to console if stated in cfg file
we have a flag stating the log level in the discovery-client.yaml.
it is defaulted to info, which was not changed.

but till now when we changed that to debug we still didn't see the debug
trace.

now we will allow it, it should not affect production, only for testing
perposes, in case we want to inhance the log level.