nvme: fix identify against wrong port
The identify implementation was a little naive in assuming that it
will get back a valid response. When running `nvme discover` against a
random port, we would crash here with a nil pointer exception. This
fix is probably needed in plenty of other places, but let's fix this
one first.

Issue: LBM1-36790

Signed-off-by: Muli Ben-Yehuda <muli@lightbitslabs.com>