serve: avoid stuck channels in reconnectToCluster()
There is a deadlock between s.aggregateChan and conn.AENChan which makes
the discovery-client a waist of space.
This change is a result of figuring out workaround with minimal
changes.

When we hit a KATO timeout, we use the conn.AENChan to notify the main
loop that we would like to create a new connection. However, this same
channel is also used for "proper" AEN events and when these two happen
in short succession, conn.AENChan blocks and the loop is stuck.
Adding a timeout such that we drop either the KATO or the AEN makes this
loop unstuck and basically drop the event. This is not enough yet. The
next blocking behaviour is in multiplexConnection() where we are
blocking on writing to s.aggregateChan. This channel is used to execute
either a discovery or reconnect (depending on payload) but since this
channel is unbuffered, this continuation too conn.AENChan might block
and so we are stuck further down the line.
To avoid this, we make s.aggregateChan buffered, giving it room for two
events.

There are other ways to do achieve the same. One such approach is to
buffer the conn.AENChan channel. This results in AEN events in quick
succession resulting in many Discovery calls which then get disconnected
right away due to current design/implementation and has other
ramifications like synchronizing access to the connection ID table.
Lastly, there is bug (LBM1-37544) where in the shutdown path we have a
use after free. Fixing this is not considered important right now since
systemd will restart it for us.

We need to ensure that events are handled and if there is a read error
due timeout or invalid response. When the service is started we use
iptables to drop packets with port 8009 and a source address of whatever
we happen to connect to. Since this connection is pseudo random so you
have to read the logs Then create volumes on all nodes (9) and validate
that we are connected.

Issue: LBM1-37388

Signed-off-by: jeffry-lb <jeffry@lightbitslabs.com>