build(Makefile): fix image url calculation
we want the default image path to be:

//:version

but we want to allow a user to override the different parts. this means
that:

default values are: DOCKER_REGISTRY==docker.lightbits.com
ORG=lightos-csi
NAME=lb-nvme-discovery-client

a user can provide a different REGISTRY or a REGISTRY/ORG
and the image name will be override accordingly.