discovery-client.yaml: comment out cores: [0]
Comment out the default specification of CPU affinity to core 0 in
discovery-client.yaml, to allow more flexible invocation even in systems
where for some reason core 0 is offline or otherwise unavailable.

It is still possible to uncomment the line and specify a list of ID's
for CPU affinity, but if the line is missing or the list is empty, no
taskset command is invoked.

Issue: LBM1-37901
Signed-off-by: Dori Polotsky <dorip@lightbitslabs.com>