config: add effective hostid to avoid connection errors
When a host is connected to the cluster and a new hostnqn, for a new
cluster member is added where, the hostid is different from the current
hostid the connection will fail.

The reason for this is that the hostnqn and the hostid must be a 1:1
mapping which is enforced by the kernel client driver since commit
ae8bd606e09bbdb2607c8249872cc2aeaf2fcc72

If/when this happens, new volumes that appear on the entry with
mismatching hostid will never appear on the host since the kernel will
reject the connection.

The solution is to scan the current hostnqn with fabric connections and
override the configured hostid with the effective hostid, being the
hostid the client is currently connected with.

When all controllers of the hostnqn are disconnected, subsequent
connects will use the configured hostid.

Issue: LBM1-40087

Signed-off-by: jeffry-lb <jeffry@lightbitslabs.com>